+++ R/biproportional.R
@@ -23,7 +23,7 @@
 #'   vector that denotes for each party/row whether they reached the quorum (i.e. are
 #'   eligible for seats). The easiest way to do this is via [quorum_any()] or
 #'   [quorum_all()], see examples. Alternatively you can pass a precalculated logical
-#'   vector. No quorum is applied if parameter is missing or `NULL`.
+#'   vector. No quorum is applied if this parameter is missing or `NULL`.
 #' @param method Defines which method is used to assign seats. The following methods are
 #'   recommended:
 #'   \itemize{
@@ -39,6 +39,7 @@
 #'   It is also possible to use any divisor method name listed in [proporz()]. If you want to
 #'   use a different method for the upper and lower apportionment, provide a list with two
 #'   entries.
+#' @param ... ignored (available for backwards compatibility)
 #'
 #' @note The iterative process in the lower apportionment is only guaranteed to terminate
 #'   with the default Sainte-Laguë/Webster method.
@@ -68,13 +69,16 @@
 biproporz = function(votes_matrix,
                      district_seats,
                      quorum,
-                     use_list_votes = TRUE,
-                     method = "round") {
+                     weight_votes = TRUE,
+                     method = "round",
+                     ...) {
+    weight_votes <- catch_deprecated_use_list_votes(weight_votes, ...)
+
     # check parameters
-    .votes_matrix.name = deparse(substitute(votes_matrix))
-    .district_seats.name = deparse(substitute(district_seats))
-    votes_matrix <- prep_votes_matrix(votes_matrix, .votes_matrix.name)
-    district_seats <- prep_district_seats(district_seats, votes_matrix, .district_seats.name, .votes_matrix.name)
+    .vmn = deparse(substitute(votes_matrix))
+    .dsn = deparse(substitute(district_seats))
+    votes_matrix <- prep_votes_matrix(votes_matrix, .vmn)
+    district_seats <- prep_district_seats(district_seats, votes_matrix, .dsn, .vmn)
     method <- prep_method(method)
 
     # quorum
@@ -83,7 +87,7 @@ biproporz = function(votes_matrix,
     }
 
     # upper apportionment (Oberzuteilung)
-    upp_app = upper_apportionment(votes_matrix, district_seats, use_list_votes, method[[1]])
+    upp_app = upper_apportionment(votes_matrix, district_seats, weight_votes, method[[1]])
 
     # lower apportionment (Unterzuteilung)
     seats_matrix = lower_apportionment(votes_matrix, upp_app$district, upp_app$party, method[[2]])
@@ -114,30 +118,31 @@ biproporz = function(votes_matrix,
 #' If you want to use other apportion methods than Sainte-Laguë use [biproporz()].
 #'
 #' @param votes_df data.frame (long format) with 3 columns (actual colnames can differ):
-#'                 \itemize{
-#'                   \item party id/name
-#'                   \item district id/name
-#'                   \item votes
+#'                 \enumerate{
+#'                   \item party id/name (character)
+#'                   \item district id/name (character)
+#'                   \item votes (numeric)
 #'                   }
 #' @param district_seats_df data.frame with 2 columns (actual colnames can differ):
-#'                          \itemize{
-#'                            \item district id/name
-#'                            \item number of seats for a district
+#'                          \enumerate{
+#'                            \item district id/name (character)
+#'                            \item number of seats for a district (numeric)
 #'                          }
 #' @inheritParams biproporz
 #' @param new_seats_col name of the new column
-#' @param use_list_votes By default (`TRUE`) it's assumed that each voter in a district has
+#' @param weight_votes By default (`TRUE`) it is assumed that each voter in a district has
 #'   as many votes as there are seats in a district. Set to `FALSE` if `votes_df` shows the
-#'   number of voters (e.g. they can only vote for one party).
+#'   number of _voters_ (e.g. because they can only vote for one party).
 #' @param winner_take_one Set to `TRUE` if the party that got the most votes in a district
 #'   must get _at least_ one seat ('Majorzbedingung') in this district. This only applies if
-#'   they are entitled to a seat in the upper apportionment. Default is `FALSE`.
+#'   the district winnig party is entitled to a seat in the upper apportionment. Default is `FALSE`.
+#' @param ... ignored (available for backwards compatibility)
 #'
 #' @seealso This function calls [biproporz()] after preparing the input data.
 #'
-#' @returns A data.frame like `votes_df` with a new column denoting the number seats per
-#'   party and district. Party and district divisors stored in attributes in attributes
-#'   (hidden from print, see [get_divisors()]).
+#' @returns A data.frame like `votes_df` with a new column denoting the number of seats per
+#'   party and district. Party and district divisors stored in attributes (hidden from print,
+#'   see [get_divisors()]). An ungrouped tibble is returned if `votes_df` is a tibble.
 #'
 #' @examples
 #' # Zug 2018
@@ -155,17 +160,19 @@ biproporz = function(votes_matrix,
 #' finland19_result = pukelsheim(finland2019$votes_df,
 #'                              finland2019$district_seats_df,
 #'                              new_seats_col = "mandates",
-#'                              use_list_votes = FALSE)
+#'                              weight_votes = FALSE)
 #' tail(finland19_result[order(finland19_result$mandates),])
 #'
 #' @export
 pukelsheim = function(votes_df, district_seats_df,
                       quorum,
                       new_seats_col = "seats",
-                      use_list_votes = TRUE,
-                      winner_take_one = FALSE) {
+                      weight_votes = TRUE,
+                      winner_take_one = FALSE,
+                      ...) {
+    weight_votes <- catch_deprecated_use_list_votes(weight_votes, ...)
 
-    check_params.pukelsheim(votes_df, district_seats_df, new_seats_col, use_list_votes, winner_take_one,
+    check_params.pukelsheim(votes_df, district_seats_df, new_seats_col, weight_votes, winner_take_one,
                             deparse(substitute(votes_df)), deparse(substitute(district_seats_df)))
 
     # Create votes matrix
@@ -178,17 +185,23 @@ pukelsheim = function(votes_df, district_seats_df,
     method = ifelse(winner_take_one, "wto", "round")
     m = biproporz(votes_matrix, district_seats,
                   quorum = quorum,
-                  use_list_votes = use_list_votes,
+                  weight_votes = weight_votes,
                   method = method)
     seats_df = pivot_to_df(m, new_seats_col)
 
     # join with original table
     assert(nrow(votes_df) <= nrow(seats_df))
-    return_df = merge(votes_df,
-                      seats_df,
+    return_df = merge(as.data.frame(votes_df),
+                      as.data.frame(seats_df),
                       sort = FALSE,
                       by = colnames(votes_df)[1:2])
-    class(return_df) <- class(votes_df)
+
+    # store divisors
     attributes(return_df)$divisors <- attributes(m)$divisors
+
+    # as_tibble, remove groups
+    if(inherits(votes_df, "tbl_df")) {
+        return_df <- .as_tibble(return_df)
+    }
     return(return_df)
 }